GYM PRJCT/ai_fitness_assistant/requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
plotly==5.18.0
Pillow==10.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
openpyxl==3.1.2
bcrypt==4.1.2
